Add back changes for frozen objects
